♻️ refactor: initInputBar 리팩토링
- 하나의 함수로 통합하고 동작을 분리하여 부분 동작을 export 하도록 리팩토링
